Moved code from misc.c to cksum_utils.c
