+++ lib/Kbuild.sub
@@ -25,6 +25,8 @@ m0tr_objects += lib/assert.o \
                   lib/m0lib.o \
                   lib/memory.o \
                   lib/misc.o \
+                  lib/cksum.o \
+                  lib/cksum_utils.o \
                   lib/mutex.o \
                   lib/queue.o \
                   lib/protocol_xc.o \
